fix/HU-05-06-07:top section, middle section and bottom section on one page.
